requirements.txt
Cython==0.29.9
matplotlib==3.0.3
mmcv==0.2.8
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.1
